remove all blue links on apple devices
Added an apple 'hack' which removes all default formatting styles seen on iOS devices and Apple Mail: a[x-apple-data-detectors]. A former employee at my company found this by viewing the source code of an email on his phone. It's 99% if not all the way bulletproof. I even made a website for it to help others - like this boilerplate! http://removebluelinks.com/